+++ src/shared/actions/marketData.js
@@ -78,7 +78,11 @@ export const getMarketData = async (dispatch) => {
 
             // Set rates statistics
             if (typeof marketData.rates === 'object') {
-                const orderedRates = reduce(keys(marketData.rates).sort(), (result, key) => (result[key] = marketData.rates[key], result), {});
+                const orderedRates = reduce(
+                    keys(marketData.rates).sort(),
+                    (result, key) => ((result[key] = marketData.rates[key]), result),
+                    {},
+                );
 
                 dispatch({
                     type: MarketDataActionTypes.SET_RATES_DATA,
@@ -95,14 +99,4 @@ export const getMarketData = async (dispatch) => {
     return fetched;
 };
 
-export function changeCurrency(currency) {
-    return (dispatch) => {
-        dispatch(setCurrency(currency));
-    };
-}
-
-export function changeTimeframe(currency, timeframe) {
-    return (dispatch) => {
-        dispatch(setTimeframe(timeframe));
-    };
-}
+export default getMarketData;
